pyproject.toml
[tool.pdm]
distribution = true

[tool.pdm.build]
includes = ["nonebot_plugin_fishspeech_tts"]
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[project]
name = "nonebot-plugin-fishspeech-tts"
version = "0.3.9"
description = "小样本TTS,通过调用在线或本地api发送TTS语音"
authors = [
    {name = "Cvandia",email = "106718176+Cvandia@users.noreply.github.com"},
]
dependencies = [
    "ormsgpack<2.0.0,>=1.7.0",
    "httpx>=0.27.0",
    "nonebot2>=2.2.0",
    "nonebot-plugin-alconna>=0.46.4",
]
requires-python = "<4.0,>=3.10"
readme = "README.md"
license = {text = "MIT"}

[dependency-groups]
dev = ["ruff>=0.7.4", "pre-commit>=4.0.1"]
